refactor: adjust the directory structure of the react node
