--- packages/g6-extension-react/src/elements/index.ts
@@ -1,1 +1,0 @@
-export * from './react';
