+++ packages/g6-extension-react/src/react-node/index.ts
