+++ packages/g6-extension-react/src/react-node/render.ts
